5_acceptance_testing

+++ app/controllers/answers_controller.rb
@@ -1,29 +1,32 @@
 class AnswersController < ApplicationController
-
-  def new; end
-
-  def edit; end
+  helper_method :question
+  helper_method :answer
 
   def create
     @answer = question.answers.new(answer_params)
+    @answer.user = current_user
     if @answer.save
-      redirect_to question
+      redirect_to question, notice: "Your answer have been successfully added"
     else
-      render :new
+      render 'questions/show'
     end
   end
 
   def update
     if answer.update(answer_params)
-      redirect_to question
+       redirect_to question, notice: "Your answer have been successfully updated"
     else
-      render :new
+      render 'questions/show'
     end
   end
 
   def destroy
-    answer.destroy
-    redirect_to question
+    if current_user.author_of?(answer)
+      answer.destroy
+      redirect_to question, notice: "Answer deleted"
+    else 
+      redirect_to question, notice: "No access to delete this answer"
+    end
   end
 
   private
@@ -32,13 +35,10 @@ class AnswersController < ApplicationController
     @answer ||= params[:id] ? Answer.find(params[:id]) : question.answers.new
   end
 
-  helper_method :answer
 
   def question
-    @question ||= Question.find(params[:question_id])
+    @question ||= params[:question_id] ? Question.find(params[:question_id]) : answer.question
   end
-
-  helper_method :question
 
   def answer_params
     params.require(:answer).permit(:body)
